% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_comparison_name}
\alias{create_comparison_name}
\title{Build a Comparison Label}
\usage{
create_comparison_name(exp, ctrl, prefix = "")
}
\arguments{
\item{exp}{Character. Name of the experimental group.}

\item{ctrl}{Character. Name of the control group.}

\item{prefix}{Character. Optional string prepended to the label. Default \code{""}.}
}
\value{
Character string of the form \code{"<prefix><exp> vs. <ctrl>"}.
}
\description{
Build a Comparison Label
}
