% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPRmd-package.R
\docType{package}
\name{DAPRmd-package}
\alias{DAPRmd}
\alias{DAPRmd-package}
\title{DAPRmd: Proteomics Core Differential Abundance Pipeline}
\description{
Differential abundance analysis for label-free quantification proteomics and phosphoproteomics data using limma. Supports protein-level and phosphopeptide-level workflows with optional GSEA and HTML report generation.
}
\author{
\strong{Maintainer}: J. Reyna \email{reynaj@vcu.edu}

}
\keyword{internal}
